# openapi.json
{"openapi":"3.0.2","info":{"title":"FundMatch","description":"Мегапроект по сбору средств. Описание добавить позже","version":"0.1.0"},"paths":{"/healthcheck/ping":{"get":{"tags":["Health check"],"summary":"Check Cart","description":"Health Check","operationId":"check_cart_healthcheck_ping_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Check Cart Healthcheck Ping Get","type":"object"}}}}}}},"/auth/login":{"post":{"tags":["Auth"],"summary":"Login","description":"1. Принимает на вход номер телефона\n2. Создает пользователя если его не существовало\n3. Генерируют одноразовый код входа\n4. Отправляет смс с кодом пользователю","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Login Auth Login Post","type":"object"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/confirm-code":{"post":{"tags":["Auth"],"summary":"Confirm Code","description":"1. Принимает на вход код подтверждения\n2. При валидном коде высылает токен авторизации","operationId":"confirm_code_auth_confirm_code_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmCodeRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Confirm Code Auth Confirm Code Post","type":"object"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/get-credentials":{"get":{"tags":["Auth"],"summary":"Get Credentials","description":"Ручка принимает токен в заголовках и возвращает данные пользователя","operationId":"get_credentials_auth_get_credentials_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"APIKeyHeader":[]}]}},"/auth/get-user":{"get":{"tags":["Auth"],"summary":"Get User","description":"Возвращает информацию о пользователе по номеру","operationId":"get_user_auth_get_user_get","parameters":[{"required":true,"schema":{"title":"Phone","type":"string"},"name":"phone","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/team/create":{"post":{"tags":["Team"],"summary":"Create Team","description":"Создание команды","operationId":"create_team_team_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamCreateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/team/{team_id}":{"get":{"tags":["Team"],"summary":"Get Team","description":"Просмотр команды","operationId":"get_team_team__team_id__get","parameters":[{"required":true,"schema":{"title":"Team Id","type":"integer"},"name":"team_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/team/invite":{"post":{"tags":["Team"],"summary":"Invite Member","description":"Пригласить участника в команду","operationId":"invite_member_team_invite_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamMemberRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/team/remove-member":{"post":{"tags":["Team"],"summary":"Remove Member","description":"Удалить члена команды","operationId":"remove_member_team_remove_member_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamMemberRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}}},"components":{"schemas":{"ConfirmCodeRequest":{"title":"ConfirmCodeRequest","type":"object","properties":{"code":{"title":"Code","type":"string","default":"491341"},"phone":{"title":"Phone","type":"string","default":"+77771119999"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"LoginRequest":{"title":"LoginRequest","type":"object","properties":{"phone":{"title":"Phone","type":"string","default":"+77771119999"}}},"TeamCreateRequest":{"title":"TeamCreateRequest","type":"object","properties":{"name":{"title":"Name","type":"string","default":"Navi"}}},"TeamCreateResponse":{"title":"TeamCreateResponse","required":["id","manager","members"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"name":{"title":"Name","type":"string","default":"Navi"},"manager":{"$ref":"#/components/schemas/UserResponse"},"members":{"title":"Members","type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}},"TeamMemberRequest":{"title":"TeamMemberRequest","required":["id"],"type":"object","properties":{"id":{"title":"Id","type":"integer"}}},"UserResponse":{"title":"UserResponse","required":["id"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"phone":{"title":"Phone","type":"string","default":"+77771119999"},"full_name":{"title":"Full Name","type":"string","default":"John Doe"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}},"securitySchemes":{"APIKeyHeader":{"type":"apiKey","in":"header","name":"Authorization"}}}}